requirements.txt
praw==7.7.1
pyspark==3.5.1
python-dotenv==1.0.1
vaderSentiment==3.3.2
pyyaml==6.0.2
pandas==2.2.2
pyarrow==17.0.0
streamlit==1.39.0
altair==5.4.1
